move(completions): into /feature
